#31: redefine meals.
- update fixtures
- add query builders

+++ src/Controller/MealController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\EatenMeal;
 use App\Entity\Meal;
+use App\Entity\User;
 use App\Form\EatenMealType;
 use App\Form\MealType;
 use App\Repository\MealRepository;
@@ -21,10 +22,20 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 class MealController extends AbstractController
 {
     #[Route('/', name: 'app_meal_index', methods: ['GET'])]
-    public function index(MealRepository $mealRepository): Response
+    public function index(Request $request, MealRepository $mealRepository): Response
     {
+        $ownMealsOnly = $request->query->get('ownMeals');
+
+        if ($ownMealsOnly === 'true') {
+            $meals = $mealRepository->findOwnMeals($this->getUser());
+        } else {
+            $meals = $mealRepository->findPublicOrOwn($this->getUser());
+            $ownMealsOnly = 'false';
+        }
+
         return $this->render('meal/index.html.twig', [
-            'meals' => $mealRepository->findAll(),
+            'meals' => $meals,
+            'ownMealsOnly' => $ownMealsOnly,
         ]);
     }
 
@@ -57,6 +68,8 @@ class MealController extends AbstractController
                 $meal->setImageFilename($newFilename);
             }
 
+            $meal->setUser($this->getUser());
+
             $entityManager->persist($meal);
             $entityManager->flush();
 
@@ -78,14 +91,46 @@ class MealController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_meal_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Meal $meal, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Meal $meal, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
     {
+        $user = $this->getUser();
+
+        assert($user instanceof User);
+        if ($meal->getUser()->getId() !== $user->getId()) {
+            $this->addFlash('danger', 'You cannot edit a meal from another user.');
+            return $this->redirectToRoute('app_meal_index');
+        }
+
         $form = $this->createForm(MealType::class, $meal);
+        /** @var FormInterface $form */
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            /** @var UploadedFile $imageFile */
+            $imageFile = $form->get('image')->getData();
+
+            if ($imageFile) {
+                $originalFilename = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME);
+                $safeFilename = $slugger->slug($originalFilename);
+                $newFilename = $safeFilename . '-' . uniqid() . '.' . $imageFile->guessExtension();
+
+
+                try {
+                    $imageFile->move(
+                        $this->getParameter('meal_image_directory'),
+                        $newFilename
+                    );
+                } catch (FileException $e) {
+                    // ... handle exception if something happens during file upload
+                }
+
+                $meal->setImageFilename($newFilename);
+            }
+
+            $entityManager->persist($meal);
             $entityManager->flush();
 
+            $this->addFlash('success', 'Meal ' . $meal->getName() . ' has been updated.');
             return $this->redirectToRoute('app_meal_index', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -98,6 +143,14 @@ class MealController extends AbstractController
     #[Route('/delete/{id}', name: 'app_meal_delete')]
     public function delete(Request $request, Meal $meal, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+
+        assert($user instanceof User);
+        if ($meal->getUser()->getId() !== $user->getId()) {
+            $this->addFlash('danger', 'You cannot delete a meal from another user.');
+            return $this->redirectToRoute('app_meal_index');
+        }
+
         $entityManager->remove($meal);
         $entityManager->flush();
